// src/enums.ts
export enum Protocol {
  http = 'http',
  https = 'https',
  socks4 = 'socks4',
  socks5 = 'socks5',
  unknown = 'unknown',
}

export enum AnonymityLevel {
  transparent = 'transparent',
  anonymous = 'anonymous',
  elite = 'elite',
  unknown = 'unknown',
}

export enum Source {
  HIDEMEIP = 'hideip.me',
  CHECKERPROXY = 'checkerproxy.net',
}